FGT indices decomposition (EXPERIMENTAL)

Estimate the Foster et al. (1984) poverty class and its components

svyfgtdec(formula, design, ...)

# S3 method for survey.design
svyfgtdec(
  formula,
  design,
  g,
  type_thresh = "abs",
  abs_thresh = NULL,
  percent = 0.6,
  quantiles = 0.5,
  na.rm = FALSE,
  thresh = FALSE,
  ...
)

# S3 method for svyrep.design
svyfgtdec(
  formula,
  design,
  g,
  type_thresh = "abs",
  abs_thresh = NULL,
  percent = 0.6,
  quantiles = 0.5,
  na.rm = FALSE,
  thresh = FALSE,
  return.replicates = FALSE,
  ...
)

# S3 method for DBIsvydesign
svyfgtdec(formula, design, ...)

Arguments

formula

a formula specifying the income variable

design

a design object of class survey.design or class svyrep.design from the survey library.

...

additional arguments. Currently not used.

g

If g=2 estimates the average squared normalised poverty gap. This function is defined for g >= 2 only,

type_thresh

type of poverty threshold. If "abs" the threshold is fixed and given the value of abs_thresh; if "relq" it is given by percent times the quantile; if "relm" it is percent times the mean.

abs_thresh

poverty threshold value if type_thresh is "abs"

percent

the multiple of the the quantile or mean used in the poverty threshold definition

quantiles

the quantile used used in the poverty threshold definition

na.rm

Should cases with missing values be dropped?

thresh

return the poverty threshold value

return.replicates

Return the replicate estimates?

Value

Object of class "cvydstat", with estimates for the FGT(g), FGT(0), FGT(1), income gap ratio and GEI(income gaps; epsilon = g) with a "var" attribute giving the variance-covariance matrix. A "statistic" attribute giving the name of the statistic.

Details

you must run the convey_prep function on your survey design object immediately after creating it with the svydesign or svrepdesign function.

Note

This function is experimental and is subject to change in later versions.

References

Oihana Aristondo, Cassilda Lasso De La vega and Ana Urrutia (2010). A new multiplicative decomposition for the Foster-Greer-Thorbecke poverty indices. Bulletin of Economic Research, Vol.62, No.3, pp. 259-267. University of Wisconsin. <doi:10.1111/j.1467-8586.2009.00320.x>

James Foster, Joel Greer and Erik Thorbecke (1984). A class of decomposable poverty measures. Econometrica, Vol.52, No.3, pp. 761-766.

Guillaume Osier (2009). Variance estimation for complex indicators of poverty and inequality. Journal of the European Survey Research Association, Vol.3, No.3, pp. 167-195, ISSN 1864-3361, URL https://ojs.ub.uni-konstanz.de/srm/article/view/369.

Jean-Claude Deville (1999). Variance estimation for complex statistics and estimators: linearization and residual techniques. Survey Methodology, 25, 193-203, URL https://www150.statcan.gc.ca/n1/en/catalogue/12-001-X19990024882.

See also

Author

Guilherme Jacob, Djalma Pessoa and Anthony Damico

Examples

library(survey)
library(laeken)
data(eusilc) ; names( eusilc ) <- tolower( names( eusilc ) )

# linearized design

des_eusilc <- svydesign( ids = ~rb030 , strata = ~db040 ,  weights = ~rb050 , data = eusilc )
des_eusilc <- convey_prep( des_eusilc )

# replicate-weighted design
des_eusilc_rep <- as.svrepdesign( des_eusilc , type = "bootstrap" )
des_eusilc_rep <- convey_prep( des_eusilc_rep )

# absolute poverty threshold
svyfgtdec(~eqincome, des_eusilc, g=2, abs_thresh=10000)
#> Warning: The svyfgtdec function is experimental and is subject to changes in later versions.
#>                     fgt2 decomposition     SE
#> fgt2                          0.016189 0.0007
#> fgt0                          0.114440 0.0027
#> fgt1                          0.032085 0.0011
#> igr                           0.280369 0.0063
#> gei(poor;epsilon=2)           0.399834 0.0132
# poverty threshold equal to arpt
svyfgtdec(~eqincome, des_eusilc, g=2, type_thresh= "relq" , thresh = TRUE)
#> Warning: The svyfgtdec function is experimental and is subject to changes in later versions.
#>                     fgt2 decomposition     SE
#> fgt2                          0.019186 0.0008
#> fgt0                          0.144442 0.0028
#> fgt1                          0.039809 0.0011
#> igr                           0.275608 0.0053
#> gei(poor;epsilon=2)           0.374324 0.0107
# poverty threshold equal to 0.6 times the mean
svyfgtdec(~eqincome, des_eusilc, g=2, type_thresh= "relm" , thresh = TRUE)
#> Warning: The svyfgtdec function is experimental and is subject to changes in later versions.
#>                     fgt2 decomposition     SE
#> fgt2                          0.023702 0.0008
#> fgt0                          0.188180 0.0028
#> fgt1                          0.051187 0.0011
#> igr                           0.272010 0.0045
#> gei(poor;epsilon=2)           0.351163 0.0089

# using svrep.design:
# absolute poverty threshold
svyfgtdec(~eqincome, des_eusilc_rep, g=2, abs_thresh=10000)
#> Warning: The svyfgtdec function is experimental and is subject to changes in later versions.
#>                     fgt2 decomposition     SE
#> fgt2                          0.016189 0.0007
#> fgt0                          0.114440 0.0027
#> fgt1                          0.032085 0.0010
#> igr                           0.280369 0.0077
#> gei(poor;epsilon=2)           0.399834 0.0158
# poverty threshold equal to arpt
svyfgtdec(~eqincome, des_eusilc_rep, g=2, type_thresh= "relq" , thresh = TRUE)
#> Warning: The svyfgtdec function is experimental and is subject to changes in later versions.
#>                     fgt2 decomposition     SE
#> fgt2                          0.019186 0.0008
#> fgt0                          0.144442 0.0030
#> fgt1                          0.039809 0.0010
#> igr                           0.275608 0.0057
#> gei(poor;epsilon=2)           0.374324 0.0107
# poverty threshold equal to 0.6 times the mean
svyfgtdec(~eqincome, des_eusilc_rep, g=2, type_thresh= "relm" , thresh = TRUE)
#> Warning: The svyfgtdec function is experimental and is subject to changes in later versions.
#>                     fgt2 decomposition     SE
#> fgt2                          0.023702 0.0008
#> fgt0                          0.188180 0.0032
#> fgt1                          0.051187 0.0011
#> igr                           0.272010 0.0041
#> gei(poor;epsilon=2)           0.351163 0.0089

if (FALSE) {

# database-backed design
library(RSQLite)
library(DBI)
dbfile <- tempfile()
conn <- dbConnect( RSQLite::SQLite() , dbfile )
dbWriteTable( conn , 'eusilc' , eusilc )

dbd_eusilc <-
  svydesign(
    ids = ~rb030 ,
    strata = ~db040 ,
    weights = ~rb050 ,
    data="eusilc",
    dbname=dbfile,
    dbtype="SQLite"
  )

dbd_eusilc <- convey_prep( dbd_eusilc )

# absolute poverty threshold
svyfgtdec(~eqincome, dbd_eusilc, g=2, abs_thresh=10000)
# poverty threshold equal to arpt
svyfgtdec(~eqincome, dbd_eusilc, g=2, type_thresh= "relq" , thresh = TRUE)
# poverty threshold equal to 0.6 times the mean
svyfgtdec(~eqincome, dbd_eusilc, g=2, type_thresh= "relm" , thresh = TRUE)

dbRemoveTable( conn , 'eusilc' )

dbDisconnect( conn , shutdown = TRUE )

}