Watts poverty index decomposition

Estimate the Watts (1968) poverty measure and its components

svywattsdec(formula, design, ...)

# S3 method for survey.design
svywattsdec(
  formula,
  design,
  type_thresh = "abs",
  abs_thresh = NULL,
  percent = 0.6,
  quantiles = 0.5,
  na.rm = FALSE,
  thresh = FALSE,
  ...
)

# S3 method for svyrep.design
svywattsdec(
  formula,
  design,
  type_thresh = "abs",
  abs_thresh = NULL,
  percent = 0.6,
  quantiles = 0.5,
  na.rm = FALSE,
  thresh = FALSE,
  return.replicates = FALSE,
  ...
)

# S3 method for DBIsvydesign
svywattsdec(formula, design, ...)

Arguments

formula

a formula specifying the income variable

design

a design object of class survey.design or class svyrep.design from the survey library.

...

additional arguments. Currently not used.

type_thresh

type of poverty threshold. If "abs" the threshold is fixed and given the value of abs_thresh; if "relq" it is given by percent times the quantile; if "relm" it is percent times the mean.

abs_thresh

poverty threshold value if type_thresh is "abs"

percent

the multiple of the the quantile or mean used in the poverty threshold definition

quantiles

the quantile used used in the poverty threshold definition

na.rm

Should cases with missing values be dropped?

thresh

return the poverty threshold value

return.replicates

Return the replicate estimates?

Value

Object of class "cvydstat", with estimates for the Watts index, FGT(0), Watts Poverty Gap Ratio, and Theil(poor incomes) with a "var" attribute giving the variance-covariance matrix. A "statistic" attribute giving the name of the statistic.

Details

you must run the convey_prep function on your survey design object immediately after creating it with the svydesign or svrepdesign function.

References

McKinley L. Blackburn (1989). Poverty measurement: an index related to a Theil measure of inequality. Journal of Business & Economic Statistics, Vol.7, No.4, pp. 475-481, URL https://amstat.tandfonline.com/doi/abs/10.1080/07350015.1989.10509760.

Satya R. Chakravarty, Joseph Deutsch and Jacques Silber (2008). On the Watts multidimensional poverty index and its decomposition. World Development, Vol.36, No.6, pp.1067-1077.

Harold W. Watts (1968). An economic definition of poverty. Institute For Research on Poverty Discussion Papers, n.5. University of Wisconsin. URL https://www.irp.wisc.edu/publications/dps/pdfs/dp568.pdf.

Guillaume Osier (2009). Variance estimation for complex indicators of poverty and inequality. Journal of the European Survey Research Association, Vol.3, No.3, pp. 167-195, ISSN 1864-3361, URL https://ojs.ub.uni-konstanz.de/srm/article/view/369.

Jean-Claude Deville (1999). Variance estimation for complex statistics and estimators: linearization and residual techniques. Survey Methodology, 25, 193-203, URL https://www150.statcan.gc.ca/n1/en/catalogue/12-001-X19990024882.

See also

Author

Guilherme Jacob, Djalma Pessoa, and Anthony Damico

Examples

library(survey)
library(laeken)
data(eusilc) ; names( eusilc ) <- tolower( names( eusilc ) )

# linearized design
des_eusilc <- svydesign( ids = ~rb030 , strata = ~db040 ,  weights = ~rb050 , data = eusilc )
des_eusilc <- convey_prep( des_eusilc )

# replicate-weighted design
des_eusilc_rep <- as.svrepdesign( des_eusilc , type = "bootstrap" )
des_eusilc_rep <- convey_prep( des_eusilc_rep )

# filter positive incomes
des_eusilc <- subset( des_eusilc , eqincome > 0 )
des_eusilc_rep <- subset( des_eusilc_rep , eqincome > 0 )

# absolute poverty threshold
svywattsdec(~eqincome, des_eusilc, abs_thresh=10000)
#>                      watts decomposition     SE
#> watts                           0.051744 0.0023
#> fgt0                            0.114257 0.0027
#> watts pov. gap ratio            0.327210 0.0087
#> theil(poor)                     0.125664 0.0101
# poverty threshold equal to arpt
svywattsdec(~eqincome, des_eusilc, type_thresh= "relq" , thresh = TRUE)
#>                      watts decomposition     SE
#> watts                           0.062404 0.0024
#> fgt0                            0.144265 0.0028
#> watts pov. gap ratio            0.320991 0.0073
#> theil(poor)                     0.111573 0.0084
# poverty threshold equal to 0.6 times the mean
svywattsdec(~eqincome, des_eusilc, type_thresh= "relm" , thresh = TRUE)
#>                      watts decomposition     SE
#> watts                           0.078038 0.0024
#> fgt0                            0.188012 0.0028
#> watts pov. gap ratio            0.316370 0.0061
#> theil(poor)                     0.098698 0.0069

#  using svrep.design:
# absolute poverty threshold
svywattsdec(~eqincome, des_eusilc_rep, abs_thresh=10000)
#>                      watts decomposition     SE
#> watts                           0.051744 0.0024
#> fgt0                            0.114257 0.0028
#> watts pov. gap ratio            0.327210 0.0093
#> theil(poor)                     0.125664 0.0091
# poverty threshold equal to arpt
svywattsdec(~eqincome, des_eusilc_rep, type_thresh= "relq" , thresh = TRUE)
#>                      watts decomposition     SE
#> watts                           0.062404 0.0026
#> fgt0                            0.144265 0.0030
#> watts pov. gap ratio            0.320991 0.0074
#> theil(poor)                     0.111573 0.0077
# poverty threshold equal to 0.6 times the mean
svywattsdec(~eqincome, des_eusilc_rep, type_thresh= "relm" , thresh = TRUE)
#>                      watts decomposition     SE
#> watts                           0.078038 0.0025
#> fgt0                            0.188012 0.0030
#> watts pov. gap ratio            0.316370 0.0070
#> theil(poor)                     0.098698 0.0064

if (FALSE) {

# database-backed design
library(RSQLite)
library(DBI)
dbfile <- tempfile()
conn <- dbConnect( RSQLite::SQLite() , dbfile )
dbWriteTable( conn , 'eusilc' , eusilc )

dbd_eusilc <-
  svydesign(
    ids = ~rb030 ,
    strata = ~db040 ,
    weights = ~rb050 ,
    data="eusilc",
    dbname=dbfile,
    dbtype="SQLite"
  )

dbd_eusilc <- convey_prep( dbd_eusilc )
dbd_eusilc <- subset( dbd_eusilc , eqincome > 0 )

# absolute poverty threshold
svywattsdec(~eqincome, dbd_eusilc, abs_thresh=10000)
# poverty threshold equal to arpt
svywattsdec(~eqincome, dbd_eusilc, type_thresh= "relq" , thresh = TRUE)
# poverty threshold equal to 0.6 times the mean
svywattsdec(~eqincome, dbd_eusilc, type_thresh= "relm" , thresh = TRUE)

dbRemoveTable( conn , 'eusilc' )

dbDisconnect( conn , shutdown = TRUE )

}